moved JwtAuthenticationException to exceptions package
