+++ src/main/java/com/cargohub/security/jwt/JwtTokenFilter.java
@@ -1,6 +1,7 @@
 package com.cargohub.security.jwt;
 
 import com.cargohub.config.RestAuthenticationEntryPoint;
+import com.cargohub.exceptions.JwtAuthenticationException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.GenericFilterBean;
