+++ src/main/java/com/cargohub/security/jwt/JwtTokenProvider.java
@@ -1,8 +1,8 @@
 package com.cargohub.security.jwt;
 
 import com.cargohub.entities.RoleEntity;
+import com.cargohub.exceptions.JwtAuthenticationException;
 import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jws;
 import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
@@ -84,7 +84,7 @@ public class JwtTokenProvider {
             }
 
             return true;
-        } catch (ExpiredJwtException e) {
+        } catch (JwtException | IllegalArgumentException e) {
             throw new JwtAuthenticationException("JWT token is expired or invalid");
         }
     }
